Teach the BGP state machine to get out of OpenSent if other end never sends an Open message. Current code simply stays in OpenSent forever. If the peer has a lower router-id, the connection collision logic will close any session(s) that he tries to open.
Ran into this situation due to a bug/misconfig in MX inline NAT which
caused BGP packets to get dropped.  This needs to be debugged further
on MX, but we should make our code more robust anyway.

Fix is to start the hold timer (4 minutes per the BGP RFC) in OpenSent
state and move to Idle state when it expires.