Partial fix for bug 1989.
There are some conditions under which we could end up replicating
an Enet route to another table.  This could happen with instances
that belong to the same VN e.g. service instances.  Hence we need
to remove the assert and just not replicate the route.

Some more elaborate checks based on OriginVn and type of instance
will be added to make this more robust.