Get bgp_update_rx_test to pass cleanly.
Problem was that the test sent both inet and inet-vpn prefixes
to a test BgpPeer, but the inet-vpn NLRI is rejected since the
address family is not configured locally on the peer.  We used
to only check the capabilities in the received Open message to
decided whether to process an NLRI earlier.  However a recent
change also checks the local configuration, thus causing the
test to fail.

Fixed by changing the hard coded default address families for
a peer from inet to (inet, inet-vpn). Tweak bgp_config_test to
match the new default address family list.