Ignore DB requests from down/deleted peers.
The old code used to add a BgpRoute from BgpTable::Input when we
received a add/change request. BgpTable::InputCommon would ignore
the request and not add a BgpPath if the request was from a peer
that's down/deleted.  This could create a scenario where we have
a BgpRoute without any BgpPaths.

Move the check for down/deleted peer from BgpTable::InputCommon
to BgpTable::Input.

Thanks to Prakash for identifying this scenario.