Test subscribe/unsubscribe to two different incarnations of instance.
The first subscribe/unsubscribe are processed after the table has been
marked as deleted but has not yet been destroyed. The second subscribe
is received before the table and instance have been destroyed. Later,
the old table and instance are destroyed and new incarnations created.
The subscribe needs to happen on the new incarnation.