Bug#2989 - agent cored at std::string::assign(std::string const&)
Fix:
For every XMPP channel in agent create a unicast sequence number.
This number will be carried with DB request. In case BGP peer has
changed in channel then the number is incremented and any request not belonging
to
the new sequence number will be ignored.
The information is being sent using data part of request.

Change-Id: I05d845dc180cc5eabdf432b67a3bdd4821da7c09