Not set decurity group on defaut project
For the SNAT implementation the ressources created (SI, VN, VM, VMI) are owned
by the default project to be hiden to the tenant and not consnume quota.
The SVC script tries to set the defaut security group of the
project to all virtual machine its created. But the default project does
not have a default security group so ignore that case.

Change-Id: I391d7f4d150f95fd107caf6433bf8cc2ac2659ef
Signed-off-by: Édouard Thuleau <edouard.thuleau@cloudwatt.com>