Fix corner case in RoutePathReplicator cleanup.
https://bugs.launchpad.net/juniperopenstack/+bug/1334063

If the last RouteTarget for which Leave is called is an export
only target, then Leave doesn't end up requesting a walk of the
VPN table.  Hence the VPN table doesn't get unregistered even
though the TableState has an empty GroupList.

Fix is to trigger a walk from Leave if the GroupList is empty.

Change-Id: Ic4ca1e7ff1a83cda87f12774abf0ce73a7931f41