Fix bug that caused double free of BgpInstanceConfig.
https://bugs.launchpad.net/juniperopenstack/+bug/1334510

BgpPeeringConfig has a back pointer to BgpInstanceConfig but does
not take a reference on the BgpInstanceConfig.  Further, the code
can delete the BgpInstanceConfig when a BgpPeeringConfig is being
deleted. If all references to BgpInstanceConfig have already gone
and it has been deleted, this results in a double free.

Fix is to keep a map of BgpPeeringConfigs in an BgpInstanceConfig
and to modify DeleteIfEmpty to check the PeeringMap.

Add logs for BgpInstanceConfig create/delete in a few places that
were overlooked previously.

Change-Id: I8579d534390b1aba7f1b14a591f51013d2ca731d