bfd: added REST API server and client
This is a rework of Bartek Wolowiec's original patch, which adds rudimentary
support for REST API as specified in [1]. There are still TODO's to make it
fully comply with the blueprint, but I believe publishing early would be
beneficial for project community.

BFD code already merged to OpenContrail suffered from coding style issues.
While working on this patch, I had to fix some of them on the way. I hope
it doesn't cloud the main intent.

[1] https://blueprints.launchpad.net/opencontrail/+spec/bfd

Signed-off-by: Bartek Wolowiec <bartek.wolowiec@codilime.com>
Signed-off-by: Mateusz Zalega <mateusz.zalega@codilime.com>
Change-Id: I17021f1d2ae91e7a41fd382ee85879f07ff2cbe1