Add router gateway support
When a gateway is set on router, the plugin dedicate a virtual network
as private network (called 'vn_snat_[ROUTER_UUID]') and start a NetNS
service instance between the private network and the public. Then
it creates a routing table with a default route that point to the
left interface of the NetNS service instance private port.

Change-Id: I53592099bda87628e90a53159e92f99777ccff10
Signed-off-by: Édouard Thuleau <edouard.thuleau@cloudwatt.com>