BgpConfig code is too aggressive in treating nodes as deleted.
BgpPeers sometimes do not come up when repeatedly doing delete
and add in the bgp_stress_test via [DELETE|ADD]_BGP_PEER events.

Root cause is that the config code does not look at IFMap nodes
that have no neighbors in the graph i.e. it considers them to
be deleted.  This can cause a problem if a bgp-router node is
added and processed before parent link to the routing-instance
has been processed.  In this case, the BgpProtocolConfig that
corresponds to the bgp-router is not created.

Fix is to not consider IFMap nodes without neighbors as deleted.

Also fix minor issue wherein the creation of a BgpPeeringConfig
would get logged as a create and an update.

Change-Id: I44c1d6de17a7a495cc18a5ddd20a05dde8b8b3fb