Increment BgpPeer flap count after setting state to IDLE.
Incrementing the flap count prior to udpating the state sometimes
causes issues for tests that first wait for the flap count to go
up and then wait for the state to reach ESTABLISHED again.  They
may see the updated flap count and old state and hence eventually
fail.

This problem was observed when feeding many CLEAR_BGP_PEER events
to a peer via the bgp_stress_test.

Change-Id: If1a7189d7b0be734a247457869aecb547e0adca1