Implement persistence for xmpp server connection flap info.
Introduce an XmppConnectionEndpoint class which keeps track
of flap count and last flap time.  The XmppServerConnection
redirects methods related to flap information to associated
XmppConnectionEndpoint object.

Note that an XmppConnectionEndpoint is not deleted till the
XmppServer is deleted.

Also make a minor change to send peer name for BGP peers in
the introspect output. There's already another field for the
address.

Change-Id: I7a1f090cb8aac2e27e59a1bc5528f5bf94db70e2