[SNAT SVC monitor] Improve scheduling on VRouter
Check if the VRouter is up and running before selected it for running a
network namespace of the SI.
Destroy and recreate the network namespace if the VRouter goes down.
This patch request the vrouter uve on the analytics API, if it receive a
HTTP 200 code, the VRouter is considered up and down otherwise.

Change-Id: Ie5c933ee3bd0a7cdac076289b4e9030db8833c82