Flap all XMPP connections if the ASN changes.
The ASN is embedded in a few extended communities such as OriginVn and
SecurityGroup.  If the ASN changes, we need to update all XMPP learnt
routes to have these extended communities based on the new ASN.  Since
an ASN change is expeceted to be a very infrequent event, we choose to
handle it by flapping all the connections instead of updating all the
routes.

Change-Id: I333decdeaaf70f5f58b84bdc85fe447bb82c8737