Bug 1348909:
For quota apis, we use vnc_api lib which reads its config from its own
config file vnc_api_lib.ini. However this file can only be read by
root and neutron runs as user name neutron. As a result, it cannot
read this file and reverts to default values of auth host/port. This
works on single node setups, but on multi node setups, if keystone is
not running locally, it fails. With this commit, we are providing a
way to pass auth parameters while creating api lib connection. With
this, our quota driver can pass these values correctly.

Change-Id: I368a7ab5dc6d4eb1ded2bd2b86877c0570a9550d