Fix the device_id and device_owner of gw
This patch fix the values returned for the
gateway port of a router.
In order to be able to get the device_id
when doing a neutron port-show this patch
also add a new link in schema between
the service instance and the logical
router.

(cherry picked from commit 8b955dcb6043655c60a03e12e9b58816d74307d8)

Conflicts:

	src/config/vnc_openstack/vnc_openstack/neutron_plugin_db.py

Change-Id: I98f8a268a3bae94e73e11ac40784009652faf848