Add a visibility property to resources
This patch add a visibility to resources
in order to hide some technical resources
from the user point of view. For example
for the SNAT we can to hide technical
resources like a network and a port.

This patch also add a property to set
if the resource has to be counted in
the quotas.

Change-Id: I77326dca9ec9de1f8b2824c2daa31b03a181812b