Fix cleanup issue in ifmap_xmpp_test
The XmppClientConnection cannot be deleted unless the ReceiveCbMap
in the XmppChannelMux is empty i.e. all receivers have unregistered.
This was a recent change to xmpp library code.

These tests did not conform to the new expectation and hence ended
up hitting an assertion when deleting the XmppVnswMockPeer.

Fix is to UnRegister before shutting down the XmppVnswMockPeer.

Change-Id: I9462e41aee4ebc52bba1db4239488cfdeaf6859a