Simplify rtarget default route handling
Don't attempt to register to all VPN tables when we receive a default
route. This avoids problems where duplicate/multiple registrations may
result in the membership manager getting stuck.

Implication is that all VPN routes added/updated after a default route
is received will be sent to the peer.  IOW, default route is effective
only if it's received before the End-Of-RIB marker is received.

Change-Id: I3b020340cf5ac403ff55847258abe91bd4558ef1