Add a user_visible property to resources
This patch add a user_visible to resources
in order to hide some technical resources
from the user point of view. For example
for the SNAT we can to hide technical
resources like a network and a port.

Resources with the property user_visible
set to false will be not count in quotas.

Change-Id: I77326dca9ec9de1f8b2824c2daa31b03a181812b