Add a user_visible property to resources
This patch add a user_visible to resources
in order to hide some technical resources
from the user point of view. For example
for the SNAT we can to hide technical
resources like a network and a port.

Resources with the property user_visible
set to false will be not count in quotas.

(cherry picked from commit 7fb6280064e6c8d8dbc7542162431203888d952e)

Conflicts:

	src/config/api-server/vnc_cfg_types.py
	src/config/svc-monitor/svc_monitor/instance_manager.py
	src/schema/vnc_cfg.xsd
Change-Id: I77326dca9ec9de1f8b2824c2daa31b03a181812b