Make evpn_prefix_test work on CentOS
Root cause was that BgpAttr::esi() returned a copy of esi_ instead
of a const reference to it.  This caused problems when copying the
esi in BgpAttr into BgpProtoPrefix in EvpnPrefix::FromProtoPrefix.
The first and last iterators that we passed to std::copy would be
pointers to different local copies of esi, causing much badness.

Change-Id: I3841e868b4d49348e940592cb4e269415ce755f2