analytics: output stream plugin system
The goal is to add a support for external .so plugins, which would be
configurable via IF-MAP database and capable of sending a real-time stream
of analytics data in any output format, with AVRO [1] plugin included in
mainline OpenContrail as a reference.

Current implementation can be configured to send out real-time Sandesh stream
in unmodified XML format. It lacks appropriate unit tests.

[1] http://avro.apache.org/

Change-Id: Id32b4db00d1cd087dab68d06ffdc615f4ecfe7af
Signed-off-by: Mateusz Zalega <mateusz.zalega@codilime.com>