Removes the user_visible default value
With a default value for this attribute
we are not able during the update operations
to know whether this attribute has to be
updated to True or if it is just the default
value. With this patch, an update without
any change for the user_visible attribute
will come with None as value.

A corresponding generator patch will be provided.

Partial-Bug: #1378376
Partial-Bug: #1377139
Change-Id: If1f38e3956bac4fe1bb56924c167d3843395f4eb