Fixes: Rabbitmq status is broken in contrail-status
contrail-status takes the service name and replaces only the first
hyphen (-) to underscore(_) to construct the corresponding UNIX sock name
for the service status query. If the service itself has more than one
hyphen, then we have to replace all hyphens with underscore.

supervisor-support-service had two hyphens and hence the issue.

Fix is to replace all the hyphens with underscores.

Change-Id: Iba2663ac8b3173056ee4111a3339e68c7d92595e
Closes-Bug:1383383