Fix potential deadlock in vnc_openstack
under certain circumstances (that alway happen in
contrail-installer + devstack environnement) keystoneclient
may authenticate before endpoints are populated in keystone.
In that case, It'll be usable to list projects, but not to
read them. Thus, it won't be reset by resync_all_projects,
but neutron api will stay stuck as projects can't be added
to vnc.

Closes-bug: #1384265
Change-Id: Iefd107604ae9f16a2b91d74c188781ef1a52394f