Support to ignore Delete request for DBEntry
Requirement:
------------
VM-Interface in agent can be added/deleted by two components
  - Thrift message
  - IFMap Config

Agent maintains a bitmask of modules that have requested addition of
an interface.

Agent should delete VM Interface only when both modules send request to
delete it. When a module requests for DBEntry to be deleted, the
corresponding bit is cleared. Delete of VM Interface is deferred till the
bitmask becomes 0

Solution:
---------
The return value of DBTable::Delete() API is changed from void to bool.
If the method returns TRUE, the DBEntry will be deleted. If the method
returns FALSE, the DBEntry is not deleted

Modified all existing Delete() methods to return TRUE

Change-Id: I42b000115301a67cc624a8b806d3b6f7d8165a4b