Add iteration facility to show route in introspect, add 40 new tests.
This limits the total number of routes we will show in each iteration.
Currently, its the min of (user-specified, 1000). This makes sure that the
backend does not run out of memory for large routing tables and also limits the
busy-time for the daemon.

This also changes the way MergeSort is working in the introspect code. The
route comparison function in bgp code sorts using numeric values via the
CompareTo function. The introspect code was doing a string compare. The output
of the 2 is not the same. This also changes the introspect code to use the
CompareTo function.

Change-Id: I9cc45181c962c3f6845174345834f6e524553c6d