Ignore deleted network during port-list as it may have disappeared.
In case of parallel tests, it is possible that during port list,
we collect a port which along with its network is deleted by the
time it is transformed to neutron representation. Ignore NoIdError
in such cases and proceed to remaining ports in the list.

Change-Id: I204e2bf9c216223ae2a918e20121950396554283
Fixes-Bug: #1384582
(cherry picked from commit 66a44bdb7304f989420129da4f909b978239bb47)