Re-establish session with ifmap server on error during publish.
Looking at logs of ifmap server, it appeared that at
1. t0 - api-server connected to ifmap server and published config-root etc.
        using port# p1
2. t1 - ifmap server reported api-server authenticated to it using port #p2
        and it deleted existing session with api-server
3. >t1 - any api-server publish to ifmap server on port# p1 fails.

event on t1 is an unresolved mystery. However same situation is created by
1. neutron net-create n1
2. ifmap-view localhost 8443 api-server api-server
3. neutron net-create n2 fails

To fix, on receiving errorResult in publish, redo newSession and re-publish.
Log on edges of failure + success. Also don't try to publish when there is
no update or delete body as that will lead to ifmap error.

Add a unit test to simulate above

Change-Id: I2468c1d86e48f1e9264e57a10805239fcd835a98
Closes-Bug: #1392132