Rabbitmq cluster connection support using Kombu client
The rabbitmq cluster can be connected with each and every
node in the cluster, or it can be connected via a load
balancer whose backend is the rabbitmq cluster.

Also, the kombu client should be able to work with a single
rabbitmq node.

This patch attempts to address the above scenarios. Following
configuration changes has to be observed.

rabbitmq_servers -
   The list of server(s) separated by comma with each server
   in the following format

   [username:[password]@]server_ip[:port]

   The rabbit_servers can also contain just one server_ip.
   The parameters in [] is optional. If not mentioned, parameters
   will substituted in the following manner

   [username] => --rabbit_user
   [password] => --rabbit_password
   [port] => --rabbit_port

rabbit_ha_mode -
   A boolean parameter. If set to True, rabbitmq queues will be
   declared/opened in high-availability mode. Defaults to False

Implements: blueprint rabbitmq-cluster-connection

Change-Id: I0564e0adb0a7a4e4ebc9b04bb853af256260858e