Update XMPP RTarget routes when BGP Identifier changes
The Nexthop of RTarget routes is set to the BGP Identifier of CN.
In order for JUNOS (and perhaps other implementations) to process
RTarget routes and trigger advertisement of relevant VPN routes,
the Nexthop must be resolved.  Hence the CN must update Nexthop
for RTarget routes when there's a change in BGP Identifier.

Change-Id: Ic6e7ff91cd8a6f95c4c5a4a1524efa7e745f15aa
Closes-Bug: 1398635