This commit makes following change: Set default cassandra connection timeout in TSocket thrift class to 3 seconds.
Reason for this change:
Currently when a pool of cassandra servers is used for Contrail Analytics service and one or more cassandra servers are down, sometimes query through UI or other clients might fail.
We see 403 http error returned by the analytics rest api. Basically issue is that
1) Query engine upon receiving query connects to cassandra servers through a socket pool.
2) Socket pool code tries connecting to one cassandra server and then moves to next if the connection attempt fails.
3) Connection failure if a server is down takes a long time. By default timeout is way more than 10s of seconds.
4) While query engine threads are stuck trying to connect to a dead cassandra server, for new queries query engine does not even acknowledge back to opserver (contrail-analytics-api).
5) Opserver (contrail-analytics-api) hence times out (it has a timeout of 10 seconds) and returns 403 error to REST api invokers.

With the change in this commit, query engine will not be stuck waiting for timeout connecting to the cassandra server.

Change-Id: Ibaedd6ba3de8473f521217ce7b46bfcee4fb9dd4
Closed-Bug: 1392575
Closes-Bug: 1398239