Merge changes to support OVS based TOR management and to support Gateway interfaces.
TOR management is based on OVS protocol. A new role called TSN (TOR services node)
is introduced for compute node. TSN handles multicast / broadcast traffic from TOR,
along with DHCP and DNS services. A new process contrail-tor-agent will be run for
every TOR that is managed. This process handles configuring TOR using OVSDB, as also
exporting the routes reported by the TOR.

Gateway Port on compute node can be configured with a subnet that is reachable on
the port.

Change-Id: I070d172e424279178d4318dc2220d6a3592ff5c1